# perf: Cache module `dir`s