@Zirak 2025-01-04
Here's the result of `pytest-benchmark compare --histogram`:
![benchmark_20250104_171648](https://github.com/user-attachments/assets/6729300f-c16e-4317-b086-12dbc98b7045)

Here's the icicle chart with `snakeviz` from before the commit:
![image](https://github.com/user-attachments/assets/4d44fb9f-9b28-4488-8991-cfb5d1236c3f)

And after:
![image](https://github.com/user-attachments/assets/1bc7e6ef-efb5-4a32-b8a0-572b44cc8601)